+++ user/user-model/src/main/java/internship/models/userModel/User.java
@@ -1,6 +1,7 @@
 package internship.models.userModel;
 
 import javax.xml.bind.annotation.XmlRootElement;
+import java.util.Set;
 
 @XmlRootElement(name = "User")
 public class User {
@@ -10,7 +11,9 @@ public class User {
 	private String patronymic;
 	private String birthday;
 	private Long passportNumber;
-	private Long income;
+	private Double income;
+
+
 
 	public long getId() {
 		return id;
@@ -60,18 +63,18 @@ public class User {
 		this.passportNumber = passportNumber;
 	}
 
-	public Long getIncome() {
+	public Double getIncome() {
 		return income;
 	}
 
-	public void setIncome(Long income) {
+	public void setIncome(Double income) {
 		this.income = income;
 	}
 
 	public User() {
 	}
 
-	public User(long id, String name, String surname, String patronymic, String birthday, Long passportNumber, Long income) {
+	public User(long id, String name, String surname, String patronymic, String birthday, Long passportNumber, Double income) {
 		this.id = id;
 		this.name = name;
 		this.surname = surname;
@@ -81,7 +84,7 @@ public class User {
 		this.income = income;
 	}
 
-	public User(String name, String surname, String patronymic, String birthday, Long passportNumber, Long income) {
+	public User(String name, String surname, String patronymic, String birthday, Long passportNumber, Double income) {
 		this.name = name;
 		this.surname = surname;
 		this.patronymic = patronymic;
